requirements.txt
python-dotenv>=1.0.0
lmstudio>=0.2.5
markdown>=3.5.1
mdformat>=0.7.22
mdformat-gfm>=0.4.1
ollama>=0.1.5
pathspec>=0.12.1
pytest>=7.4.0
pytest-cov>=4.1.0
tqdm>=4.66.1
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Development dependencies
pre-commit>=3.7.0
ruff>=0.11.0
mypy>=1.9.0
types-requests>=2.31.0
types-setuptools>=69.1.0
types-pyyaml>=6.0.12
types-toml>=0.10.8
